tranform parsing into multiple modules
